Common Rust-BPF utilities and types
