+++ programs/bpf/rust/noop/Cargo.toml
@@ -9,8 +9,10 @@ authors = ["Solana Maintainers <maintainers@solana.com>"]
 repository = "https://github.com/solana-labs/solana"
 license = "Apache-2.0"
 homepage = "https://solana.com/"
+edition = "2018"
 
 [dependencies]
+solana-sdk-rust-utils = { path = "../../../../sdk/bpf/rust-utils", version = "0.15.0" }
 # byteorder = { version = "1.3.1", default-features = false }
 # heapless = { version = "0.4.0", default-features = false }
 # byte = { version = "0.2", default-features = false }
