+++ sdk/bpf/rust-utils/Cargo.toml
@@ -1,0 +1,16 @@
+
+[package]
+name = "solana-sdk-rust-utils"
+version = "0.15.0"
+description = "Solana BPF SDK Rust Utils"
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+edition = "2018"
+
+[workspace]
+members = []
+
+
+
